move tmdb/imdb link parsing logic to MovieLookupController.cs
